+++ lib/src/Interface/isolate_native.dart
@@ -1,4 +1,1 @@
-// Re-export all needed classes and functions from dart:isolate
 export 'dart:isolate' show Isolate, SendPort, ReceivePort;
-
-// You can add additional helpers or wrappers here if needed